Extracted replaceLinks into a utility file
